Remove some caching which will not work with Rails 4.2
Rails 4.2 will have "AdequateRecord", which prevents us from
caching certain calls (unless we wanted to override find itself).

Therefore, we are trying the impact of this ahead of time in Rails 3.2 to see
what performance is like if we do not cache a few things. What we're trying
to see is the worst-case scenario. Request time will likely drop in Rails 4.2
again since "AdequateRecord" is faster than ActiveRecord in Rails 3.2 for many
scenarios which were cached previously.

Before this change, we have the following scenario:

* +all+ is cached (with the alias +cached_all+)
* +where+ is not cached
* +by_xyz+ is cached (depending on all_with_....)
* +find_by_xyz+ is cached (depending on all_with_...)
* +find+ is cached (depending on find_one_with_.../find_some_with...)
* +first+ is cached (depending on find_first_with_....)
* Constants are cached (depending on all_with_...)

After this change, we only cache constants, +all+ and +by_xyz+.

In Rails 4.2, we're going to implement +by_xyz+ by caching +where+.
We are dropping +find_by_xyz+ too since that is using "AdequateRecord" in
Rails 4.2 by default.