Remove some caching which will not work with Rails 4.2
Rails 4.2 will have "AdequateRecord", which prevents us from
caching the way we do it currently.

Before this change, we have the following scenario:

* +all+ is cached (with the alias +cached_all+)
* +where+ is not cached
* +by_xyz+ is cached (depending on all_with_....)
* +find_by_xyz+ is cached (depending on all_with_...)
* +find+ is cached (depending on find_one_with_.../find_some_with...)
* +first+ is cached (depending on find_first_with_....)
* Constants are cached (depending on all_with_...)

After this change, we still cache the same, but the implementation is
quite different. Mainly:

* +find_by_xyz+ is implemented via +by_xyz+
* +find+ is implemented via +by_id+
* +first+ is implemented via +all.first+

All usage of arguments for +all+, +find+, +first+ etc. results in
the cache not being used.

Additionally, due to the implementation change, relations were cached
previously, now they are not anymore.